them api ai cơ bản

+++ backend/src/main/java/hackathon/fridgeai/dto/AuthDto.java
@@ -29,6 +29,6 @@ public class AuthDto {
         private Long userId;
         private String name;
         private String email;
-        private Integer fridgeId;
+        private Long fridgeId;
     }
 }
